+++ packages/devtools_extensions/example/foo_devtools_extension/lib/main.dart
@@ -74,6 +74,29 @@ class _FooExtensionHomePageState extends State<FooExtensionHomePage> {
             ),
             const SizedBox(height: 48.0),
             Text('Received message from DevTools: $_message'),
+            const SizedBox(height: 48.0),
+            Row(
+              mainAxisSize: MainAxisSize.min,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                ElevatedButton(
+                  onPressed: () => extensionManager
+                      .showNotification('Yay, DevTools Extensions!'),
+                  child: const Text('Show DevTools notification'),
+                ),
+                const SizedBox(width: 16.0),
+                ElevatedButton(
+                  onPressed: () => extensionManager.showBannerMessage(
+                    key: 'example_message',
+                    type: 'warning',
+                    message:
+                        'Warning: with great power, comes great responsibility.',
+                    extensionName: 'foo_package',
+                  ),
+                  child: const Text('Show DevTools warning'),
+                ),
+              ],
+            ),
           ],
         ),
       ),