+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.3-wip
+* Add `showNotification` and `showBannerMessage` endpoints to the extensions API.
 * Add hot reload and hot restart actions to the simulated DevTools environment.
 
 ## 0.0.2